The spell “Resist Pain” DV is (Damage Value) – 6 while Chummer5 shows F – 4.
In the build summary on the right hand side of the interface, the number of Spells chosen is not always 
shown correctly. I have just now loaded a character who should have -1 of 7, but it says. 7 of 7. 
I've also noticed that it does not update correctly when you remove spells. 
However, when you add a spell, it seems it always gets it right.

When you create your character the other info tab does not show the armor information.